requirements.txt
fastapi==0.103.2
passlib[bcrypt]==1.7.4
pydantic==2.4.2
pydantic_settings==2.0.3
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
SQLAlchemy==2.0.21
uvicorn==0.23.2
psycopg2-binary==2.9.9
